requirements.txt
openai>=1.3.0
fireworks-ai
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-fireworks
sentence-transformers>=2.0.0